refactor: 按 review 意见拆分 swiglu 文件并改用混合精度容差
1. 路径迁移: examples/swiglu/ → examples/cann-bench/swiglu/
   (审核人 hedi515: 放到 example/cannbench 文件夹路径下)
2. 文件拆分: swiglu.py(kernel+adapter) + test_swiglu.py(golden+测试)
   (审核人 hedi515: kernel 单独放一个文件，test 和 golden 单独放一个文件)
3. 精度标准: assert_close(rtol=1e-2) → 混合精度容差(per-dtype atol/rtol + matched_ratio>=0.99)
   (审核人 hedi515: 精度标准采用混合精度要求)
   - float16:  atol=6.10e-5  rtol=1.95e-3
   - bfloat16: atol=9.77e-4  rtol=1.56e-2
   - float32:  atol=1.53e-5  rtol=9.77e-4
